[Gold II] Title: 최대공약수 하나 빼기, Time: 588 ms, Memory: 126848 KB -BaekjoonHub

+++ 백준/Gold/14476. 최대공약수 하나 빼기/최대공약수 하나 빼기.java
@@ -4,9 +4,9 @@ import java.util.StringTokenizer;
 
 public class Main {
 
-    public static long getGcd(long a, long b) {
+    public static int getGcd(int a, int b) {
         // a > b
-        long tmp;
+        int tmp;
         while (b != 0) {
             tmp = a % b;
             a = b;
@@ -22,26 +22,26 @@ public class Main {
 
         int N = Integer.parseInt(br.readLine());
 
-        long[] arr = new long[N];
+        int[] arr = new int[N];
         StringTokenizer st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N; i++) {
             arr[i] = Integer.parseInt(st.nextToken());
         }
 
-        long[] prefix = new long[N];
+        int[] prefix = new int[N];
         prefix[0] = arr[0];
         for (int i = 1; i < N; i++) {
             prefix[i] = getGcd(Math.max(arr[i], prefix[i - 1]), Math.min(arr[i], prefix[i - 1]));
         }
 
-        long[] suffix = new long[N];
+        int[] suffix = new int[N];
         suffix[N - 1] = arr[N - 1];
         for (int i = N - 2; i >= 0; i--) {
             suffix[i] = getGcd(Math.max(arr[i], suffix[i + 1]), Math.min(arr[i], suffix[i + 1]));
         }
 
 
-        long k, answer = -1, delK = -1, gcd;
+        int k, answer = -1, delK = -1, gcd;
         for (int i = 0; i < N; i++) {
             k = arr[i];
